fixing the throttling range to integer values

+++ src/test/java/steps/HealthCheckTest.java
@@ -107,8 +107,8 @@ public class HealthCheckTest extends BaseTest {
 
         int throttlingRequests =  requestSent - expected;
         int tolerancePercentages = 6;
-        double minValue = requestSent - (throttlingRequests * (1 + (double)tolerancePercentages / 100));
-        double maxValue = requestSent - (throttlingRequests * (1 - (double)tolerancePercentages / 100));
+        int minValue =  (int) (requestSent - (throttlingRequests * (1 + (double)tolerancePercentages / 100)));
+        int maxValue = (int) (requestSent - (throttlingRequests * (1 - (double)tolerancePercentages / 100)));
         System.out.println("total equests with status code : " + expectedResponseCode + " are ==>  " + count + " ... while the expected values are : " + (minValue) + "-" + (maxValue));
         Assert.assertTrue((count <= (maxValue)) && (count >= (minValue)));
     }
